+++ src/Database/Models/EnergyLevel.java
@@ -1,4 +1,4 @@
-package Database;
+package Database.Models;
 
 
 public class EnergyLevel {
