+++ src/Main.java
@@ -1,10 +1,10 @@
-import Database.*;
+import Database.Connection.DatabaseConnection;
+import Database.Models.AcademicGoal;
+import Database.Tables.AcademicGoalsTable;
+import Database.Tables.UsersTable;
+
 import java.sql.Connection;
-import java.sql.Date;
 import java.sql.SQLException;
-import java.sql.Timestamp;
-import java.time.LocalDate;
-import java.util.List;
 
 public class Main {
     public static void main(String[] args) {
@@ -12,6 +12,9 @@ public class Main {
             Connection connection = DatabaseConnection.getConnection();
             UsersTable users=new UsersTable(connection);
             AcademicGoalsTable academicGoalsTable = new AcademicGoalsTable(connection);
+
+            AcademicGoal academicGoal=academicGoalsTable.getById(1);
+            System.out.println(academicGoal.getUser_id());
 
 
         } catch (SQLException e) {
